Integrate worker processes with failover

+++ bagel/worker_stop.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
 pid=$(ps aux | grep "./bin/worker $1" | grep -v grep | awk -F' ' '{print $2}')
-echo "Coord killing process ${pid}"
-kill "${pid}"
-echo "Coord killed process ${pid}"
+if [ ! -z "${pid}" ]
+then
+  echo "Coord killing process ${pid}"
+  kill "${pid}"
+  echo "Coord killed process ${pid}"
+fi
